pyproject.toml
[project]
name = "herif-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py[voice]>=2.7.1",
    "google-cloud-logging>=3.6",
    "yt-dlp[default]>=2026.3.3",
    "openai>=1.43.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "instaloader>=4.12",
    "async-lru>=2.0.5",
    "axiom-py>=0.9.0",
    "psycopg[binary,pool]>=3.3.2",
    "frozendict>=2.4.7",
]

[dependency-groups]
dev = [
    "pylint>=4.0.4",
    "pylint-per-file-ignores>=3.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
